gDR: Umbrella package for R packages in the gDR suite

Package is a part of the gDR suite. It reexports functions from other packages in the gDR suite that contain critical processing functions and utilities. The vignette walks through the full processing pipeline for drug response analyses that the gDR suite offers.

Package is a part of the gDR suite. It reexports functions from other packages in the gDR suite that contain critical processing functions and utilities. The vignette walks through the full processing pipeline for drug response analyses that the gDR suite offers.

Value

package help page

package help page

Note

To learn more about functions start with help(package = "gDR")

See also

Author

Maintainer: Arkadiusz Gladki gladki.arkadiusz@gmail.com (ORCID)

Authors:

Allison Vuong vuong.allison@gene.com

Bartosz Czech (ORCID)

Marc Hafner (ORCID)

Dariusz Scigocki

Janina Smola

Sergiu Mocanu